requirements.txt
# 必要依赖
requests
pillow
pdf2image
python-dotenv
openai

# 可选依赖（用于数据库功能）
psycopg2-binary
